# src/openapi.yml
openapi: 3.0.3
info:
  title: Course Picker API
  description: Документация backend-приложения платформы рекомендаций массовых онлайн курсов
  version: 1.0.1
...
paths:
...
  /users/favourite:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - users
      summary: "Получение списка избранных курсов пользователя"
      responses:
        200:
          description: "Успешное получение списка избранного"
          content:
            application/json:
              schema:
                properties:
                  favouriteCourses:
                    type: array
                    items:
                      $ref: "#/components/schemas/CoursePreview"
        401:
          description: "Требуется корректная авторизация"
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: "Необходима авторизация"
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - users
      summary: "Добавление курса в избранные"
      parameters:
        - in: query
          name: id
          required: true
          description: "Id курса для добавления в избранные"
          schema:
            $ref: "#/components/schemas/CourseId"
      responses:
        201:
          description: "Успешное добавление курса в избранное"
          content:
            application/json:
              schema:
                properties:
                  favouriteCourses:
                    type: array
                    items:
                      type: string
                    example: [ "openedu_2", "openedu_1" ]
        401:
          description: "Требуется корректная авторизация"
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: "Необходима авторизация"
        404:
          description: "Курс не найден"
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: "Не найден курс по переданному id"
    delete:
      security:
        - bearerAuth: [ ]
      tags:
        - users
      summary: "Удаление курса из избранного"
      parameters:
        - in: query
          name: id
          required: true
          description: "Id курса для удаления из избранного"
          schema:
            $ref: "#/components/schemas/CourseId"
      responses:
        200:
          description: "Успешное удаление курса из избранного"
          content:
            application/json:
              schema:
                properties:
                  favouriteCourses:
                    type: array
                    items:
                      type: string
                    example: [ "openedu_1" ]
        401:
          description: "Требуется корректная авторизация"
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: "Необходима авторизация"
  /users/viewed:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - users
      summary: "Получение списка просмотренных курсов пользователя"
      responses:
        200:
          description: "Успешное получение списка просмотренных"
          content:
            application/json:
              schema:
                properties:
                  viewedCourses:
                    type: array
                    items:
                      $ref: "#/components/schemas/CoursePreview"
        401:
          description: "Требуется корректная авторизация"
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: "Необходима авторизация"
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - users
      summary: "Добавление курса в просмотренные"
      parameters:
        - in: query
          name: id
          required: true
          description: "Id курса для добавления в просмотренные"
          schema:
            $ref: "#/components/schemas/CourseId"
      responses:
        201:
          description: "Успешное добавление курса в просмотренные"
          content:
            application/json:
              schema:
                properties:
                 viewedCourses:
                    type: array
                    items:
                      type: string
                    example: [ "openedu_2", "openedu_1" ]
        401:
          description: "Требуется корректная авторизация"
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: "Необходима авторизация"
        404:
          description: "Курс не найден"
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: "Не найден курс по переданному id"
    delete:
      security:
        - bearerAuth: [ ]
      tags:
        - users
      summary: "Удаление курса из просмотренного"
      parameters:
        - in: query
          name: id
          required: true
          description: "Id курса для удаления из просмотренного"
          schema:
            $ref: "#/components/schemas/CourseId"
      responses:
        200:
          description: "Успешное удаление курса из просмотренного"
          content:
            application/json:
              schema:
                properties:
                  viewedCourses:
                    type: array
                    items:
                      type: string
                    example: [ "openedu_1" ]
        401:
          description: "Требуется корректная авторизация"
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: "Необходима авторизация"
components:
  schemas:
...
    Category:
      description: "Категория курсов"
      type: object
      properties:
        id:
          description: "Уникальный id категории"
          type: integer
          example: 1
        name:
          type: object
          properties:
            ru:
              description: "Русское название категории"
              type: string
              example: "Бизнес"
            en:
              description: "Английское название категории"
              type: string
              example: "Business"
    CourseLanguages:
      description: "Массив языков курсов"
      type: array
      items:
        type: string
      example: ["ru"]
    CourseId:
      description: "Уникальный id курса"
      type: string
      example: openedu_2
    CourseName:
      description: "Название курса"
      type: string
      example: "Веб-программирование"
...
    CourseShortDescription:
      description: "Короткое описание курса"
      type: string
      example: "Короткое описание курса"
    CourseLink:
      description: "Ссылка на курс"
      type: string
      example: "https://openedu.ru/course/ITMOUniversity/WEBDEV/"
    CoursePreviewImageLink:
      description: "Ссылка на картинку с превью курса"
      type: string
      example: "https://cdn.openedu.ru/f1367c/CACHE/images/cover/itmo_web_s/71c221bca9f9ad3ad57b257ae0ede1e2.png"
    CourseRating:
      description: "Рейтинг курса"
      type: object
      properties:
        external:
          description: "Внешний рейтинг курса (на сайте поставщика)"
          type: object
          properties:
            averageScore:
              type: number
              example: 4.52
            countReviews:
              type: integer
              example: 21
        internal:
          description: "Внутренний рейтинг курса"
          type: object
          properties:
            averageScore:
              type: number
              example: 4.75
            countReviews:
              type: integer
              example: 5
    CourseVendor:
      description: "Платформа-поставщик курса"
      type: object
      properties:
        id:
          description: "Id поставщика"
          type: string
          example: "openedu"
        name:
          description: "Название поставщика"
          type: string
          example: "OpenEdu"
        link:
          description: "Ссылка на главную страницу поставщика"
          type: string
          example: "https://openedu.ru/"
        icon:
          description: "Ссылка на логотип поставщика"
          type: string
          example: "https://mooc.ij.je/public_files/img/vendors/openedu.png"
    CourseAuthorName:
      description: "Название/имя автора курса"
      type: string
      example: "Университет ИТМО"
    CourseAuthorLink:
      description: "Ссылка на автора курса"
      type: string
      example: "https://openedu.ru/university/ITMOUniversity/"
    CourseAuthorIcon:
      description: "Ссылка на картинку автора курса"
      type: string
      example: "https://cdn.openedu.ru/f1367c/university-icon/itmo_small_white_rus_100_100.png"
    CourseAuthor:
      description: "Схема объекта автора курса"
      type: object
      properties:
        name:
          $ref: "#/components/schemas/CourseAuthorName"
        link:
          $ref: "#/components/schemas/CourseAuthorLink"
        icon:
          $ref: "#/components/schemas/CourseAuthorIcon"
    CourseDuration:
      description: "Длительность курса"
      type: string
      example: "10w"
    CoursePrice:
      description: "Цена курса"
      type: object
      properties:
        amount:
          description: "Цена курса"
          type: number
          example: 1200.50
        currency:
          description: "Валюта, в которой указана цена курса"
          type: string
          example: "RUB"
    CoursePreview:
      description: "Схема для описания курса"
      type: object
      properties:
        courseLanguages:
          $ref: "#/components/schemas/CourseLanguages"
        id:
          $ref: "#/components/schemas/CourseId"
        courseName:
          $ref: "#/components/schemas/CourseName"
        shortDescription:
          $ref: "#/components/schemas/CourseShortDescription"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        link:
          $ref: "#/components/schemas/CourseLink"
        previewImageLink:
          $ref: "#/components/schemas/CoursePreviewImageLink"
        rating:
          $ref: "#/components/schemas/CourseRating"
        vendor:
          $ref: "#/components/schemas/CourseVendor"
        author:
          $ref: "#/components/schemas/CourseAuthor"
        duration:
          $ref: "#/components/schemas/CourseDuration"
        price:
          $ref: "#/components/schemas/CoursePrice"
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT